Changer support float

+++ lib/acts_as_votable/votable.rb
@@ -91,7 +91,7 @@ module ActsAsVotable
       vote.vote_flag = votable_words.meaning_of(options[:vote])
 
       #Allowing for a vote_weight to be associated with every vote. Could change with every voter object
-      vote.vote_weight = (options[:vote_weight].to_i if options[:vote_weight].present?) || 1
+      vote.vote_weight = (options[:vote_weight].to_f if options[:vote_weight].present?) || 1
 
       ActiveRecord::Base.transaction do
         self.vote_registered = false
